anthurium_api/anthurium_api/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
